refactor: ViewPager를 Activity가 아닌 Fragment로 이동
