+++ app/src/main/java/com/nbc/two_of_us/presentation/MainActivity.kt
@@ -1,10 +1,8 @@
 package com.nbc.two_of_us.presentation
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import com.google.android.material.tabs.TabLayoutMediator
+import androidx.appcompat.app.AppCompatActivity
 import com.nbc.two_of_us.databinding.ActivityMainBinding
-import com.nbc.two_of_us.presentation.model.TabType
 
 class MainActivity : AppCompatActivity() {
 
@@ -12,24 +10,8 @@ class MainActivity : AppCompatActivity() {
         ActivityMainBinding.inflate(layoutInflater)
     }
 
-    private val adapter: ContactViewPagerAdapter = ContactViewPagerAdapter(this)
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
-
-        setViewPager()
-        setTabLayout()
-    }
-
-    private fun setViewPager() = with(binding) {
-        vp.adapter = adapter
-    }
-
-    private fun setTabLayout() = with(binding) {
-        TabLayoutMediator(tabLayout, vp) { tab, position ->
-            val tabType = TabType.from(position)
-            tab.text = getString(tabType.tabName)
-        }.attach()
     }
 }
