+++ app/src/main/java/com/nbc/two_of_us/presentation/viewpager/ContactViewPagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.nbc.two_of_us.presentation
+package com.nbc.two_of_us.presentation.viewpager
 
 import android.net.Uri
 import android.os.Bundle
@@ -10,7 +10,7 @@ import com.nbc.two_of_us.presentation.contact.ContactListFragment
 import com.nbc.two_of_us.presentation.contact_detail.ContactDetailFragment
 import com.nbc.two_of_us.presentation.model.TabType
 
-class ContactViewPagerAdapter(fa: FragmentActivity): FragmentStateAdapter(fa) {
+class ContactViewPagerAdapter(fragment: Fragment): FragmentStateAdapter(fragment) {
     override fun getItemCount(): Int = TabType.entries.size
 
     override fun createFragment(position: Int): Fragment {
